Changed Int to BigDecimal for estimatedAmountOfTaxInTheNext12Months and removed the TODO

+++ app/uk/gov/hmrc/softdrinksindustrylevystub/models/CreateSubscription.scala
@@ -133,7 +133,7 @@ case class CreateSubscriptionRequest (
   primaryPerson: ContactDetails,
   softDrinksIndustryLevyDetails: LevyDetails,
   sdilActivity: LitresProduced,
-  estimatedAmountOfTaxInTheNext12Months: Option[Int], // TODO change this to BigDecimal
+  estimatedAmountOfTaxInTheNext12Months: Option[BigDecimal],
   taxObligationStartDate: Date,
   bankDetails: BankDetails,
   sites: List[Site]
